Rename debugview package to ui.debugview
