--- debugview/src/main/kotlin/com/revenuecat/purchases/debugview/models/SettingGroupState.kt
@@ -1,3 +1,0 @@
-package com.revenuecat.purchases.debugview.models
-
-internal data class SettingGroupState(val title: String, val settings: List<SettingState>)
