--- debugview/src/release/kotlin/com/revenuecat/purchases/debugview/InternalDebugRevenueCatBottomSheet.kt
@@ -1,14 +1,0 @@
-package com.revenuecat.purchases.debugview
-
-import androidx.compose.runtime.Composable
-import com.revenuecat.purchases.PurchasesTransactionException
-import com.revenuecat.purchases.models.StoreTransaction
-
-@Suppress("EmptyFunctionBlock", "UnusedParameter")
-@Composable
-internal fun InternalDebugRevenueCatBottomSheet(
-    onPurchaseCompleted: (StoreTransaction) -> Unit,
-    onPurchaseErrored: (PurchasesTransactionException) -> Unit,
-    isVisible: Boolean = false,
-    onDismissCallback: (() -> Unit)? = null,
-) {}
