+++ settings.gradle
@@ -11,4 +11,4 @@ include ':integration-tests'
 include ':purchases'
 include 'examples:purchase-tester'
 include ':api-tester'
-include ':debugview'
+include 'ui:debugview'
