+++ ui/debugview/proguard-rules.pro
