+++ ui/debugview/src/main/AndroidManifest.xml
