+++ ui/debugview/src/main/kotlin/com/revenuecat/purchases/ui/debugview/models/SettingGroupState.kt
@@ -1,0 +1,3 @@
+package com.revenuecat.purchases.ui.debugview.models
+
+internal data class SettingGroupState(val title: String, val settings: List<SettingState>)
